Create Snakegame
Succsessful attempt at recreating Snake. i just used text to create the snake and food. Borders are solid and don't wrap, so be careful around the edges. There are still things im not happy about in this release with a lot of quality of life improvements to be made, but nevertheless it is still the game Snake